PlatformModuleType
